+++ packData/packData.mjs
@@ -18,7 +18,8 @@ let packs = [
     'cpr-summon-features',
     'cpr-summons',
     'cpr-spells-2024',
-    'cpr-class-features-2024'
+    'cpr-class-features-2024',
+    'cpr-items-2024'
 ];
 for (let i of packs) {
     await compilePack('./packData/' + i, './packs/' + i, {'log': true});